+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 #### Removed
 -->
 
+### [1.5.0]
+#### Added
+* File listing methods added based on given path, searchPattern and SearchOption. 
+* Folder listing methods added based on given path, searchPattern and SearchOption.
 ### [1.4.0]
 #### Removed
 * `TodayString`, `NowString` and `LongDayString` are removed as `NamingFormat`.